requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pillow
python-dotenv
requests